backend/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
# mariadb==1.1.10
MarkupSafe==2.1.5
packaging==24.0
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.2
flask-sqlalchemy==3.1.1
pyjwt==2.9.0
